Extract client methods
